feature: 분야 선택 아코디언 메뉴 컴포넌트 분리
