+++ src/components/monthlyExhibition/ContentsSection2.jsx
@@ -1,8 +1,9 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useCallback, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 import { ReactComponent as WhiteArrowIcon } from "../../assets/icons/white-arrow.svg";
-import { ReactComponent as ExpandIcon } from "../../assets/icons/expand.svg";
+
+import CategorySelectMenu from "./CategorySelectMenu";
 
 const ContentsSection2 = () => {
   const navigate = useNavigate();
@@ -25,7 +26,7 @@ const ContentsSection2 = () => {
 
   // 확장 버튼 클릭 여부 state
   const [onClickBtn, setClickBtn] = useState(false);
-  console.log(onClickBtn);
+
   return (
     <div className="section2-container">
       <WhiteArrowIcon
@@ -36,60 +37,14 @@ const ContentsSection2 = () => {
       />
       <div className="list-wrap">
         <div className="title">이달의 전시회</div>
-        <div className="category-wrap">
-          <div className="sub-title">분야</div>
-          <div
-            className="interests-wrap"
-            style={{ height: `${onClickBtn ? "max-content" : "2vw"}` }}
-          >
-            {interestsList.map((interest) => (
-              <div
-                key={interest}
-                className="interest-btn"
-                onClick={handleClick}
-              >
-                {interest}
-              </div>
-            ))}
-          </div>
-          <ExpandIcon
-            className="expand-btn"
-            style={{
-              transform: `${onClickBtn ? "rotate(180deg)" : "rotate(0deg)"}`,
-            }}
-            onClick={() => setClickBtn(!onClickBtn)}
-          />
-        </div>
+        <CategorySelectMenu
+          onClickBtn={onClickBtn}
+          setClickBtn={setClickBtn}
+          handleClick={handleClick}
+        />
       </div>
     </div>
   );
 };
 
 export default ContentsSection2;
-
-const interestsList = [
-  "예술",
-  "문화",
-  "다양성",
-  "교육",
-  "음식 및 음료",
-  "여행",
-  "자연",
-  "스토리",
-  "미디어",
-  "플랫폼",
-  "미래지향",
-  "인터랙티브",
-  "기념적인",
-  "감성적인",
-  "자연친화적인",
-  "공익적인",
-  "지속가능성",
-  "기술",
-  "우주",
-  "사랑",
-  "게임",
-  "캐릭터",
-  "사진",
-  "취미",
-];
